refactor(interfaces): Separate type-map into files
