+++ modules/interfaces/src/index.ts
@@ -1,12 +1,15 @@
 export * from "./action";
+export * from "./auth-command-map";
 export * from "./bound-request-response";
 export * from "./client";
 export * from "./command-result";
 export * from "./command";
 export * from "./custom-command-definition";
+export * from "./custom-map";
 export * from "./custom-query-definition";
 export * from "./db-client";
 export * from "./entity-definition";
+export * from "./entity-rest-info-map";
 export * from "./entity-state";
 export * from "./entity";
 export * from "./error";
